Moving code around to help my understanding. file names most probably wrong right now but will update when understand it better
